test(scroll-area): recategorize and clean up browser tests
