+++ packages/react/src/components/scroll-area/scroll-area.test.browser.tsx
@@ -1,14 +1,6 @@
 import type { FC } from "react"
-import { a11y, page, render, renderHook } from "#test/browser"
+import { page, render } from "#test/browser"
 import { ScrollArea } from "."
-import { useScrollArea } from "./use-scroll-area"
-
-const setScrollTop = (el: HTMLElement, value: number) => {
-  Object.defineProperty(el, "scrollTop", {
-    configurable: true,
-    value,
-  })
-}
 
 const TestContent: FC = () => {
   return (
@@ -30,90 +22,11 @@ const TestContent: FC = () => {
       <p>
         ピッコロ（マジュニア）との闘いから約5年後、息子の孫悟飯を儲けて平和な日々を過ごしていた悟空のもとに、実兄・ラディッツが宇宙より来襲し、自分が惑星ベジータの戦闘民族・サイヤ人であることを知らされる。さらわれた孫悟飯を助けるため悟空は宿敵ピッコロと手を組み、自らの命と引き換えにラディッツを倒すが、約1年後にはさらに強力なサイヤ人たちがドラゴンボールを求めて地球に来襲することを知る。悟空はドラゴンボールによって生き返るまでの間、あの世の界王の下で修業し、仲間と共に地球に強襲したサイヤ人の戦士・ナッパと王子・ベジータを迎え撃つ。悟空は修行により増した力でナッパを一蹴し、ベジータと決闘。仲間の協力もあり、何とか辛勝し撤退させるが、多くの仲間を失うとともに、ピッコロの戦死により彼と一心同体であった神も死亡し、地球のドラゴンボールも消滅する。
       </p>
-
-      <h1>フリーザ編</h1>
-
-      <p>
-        地球の神と殺された仲間たちを甦らせるため、重傷で入院中の悟空に代わり、悟飯、クリリン、ブルマの3人が神とピッコロの故郷であるナメック星へ向かう。だが、そこには地球で闘ったベジータや、界王すら畏怖する宇宙の帝王・フリーザとその一味が不老不死を求めて来襲し、ナメック星人を虐殺しながらドラゴンボールを略奪していた。悟飯たちはベジータ、フリーザ一味とのドラゴンボールをめぐる三つ巴の攻防の末、後から到着した悟空とナメック星人たちの協力を得てナメック星の神龍・ポルンガを呼び出し、ピッコロと地球のドラゴンボールを復活させる。出し抜かれて願いが叶えられなかったフリーザは激怒し、一行は対決を強いられる。フリーザの持つ圧倒的な力の前にはベジータやピッコロ、悟空すら歯が立たず仲間たちが次々と命を落としていった。怒りを爆発させた悟空は伝説の戦士・超（スーパー）サイヤ人へと覚醒。フルパワーを解放したフリーザに勝利する。ポルンガによって地球に帰還した悟飯たちは復活したドラゴンボールによりサイヤ人やフリーザ一味に殺された人々を蘇生させた。一方の悟空も爆発するナメック星を辛くも脱出、ヤードラット星に漂着し一命を取り留めた。
-      </p>
-
-      <h1>人造人間・セル編</h1>
-
-      <p>
-        ナメック星での闘いから約1年後、密かに生き延びていたフリーザとその一味が地球を襲撃するが、謎の超サイヤ人によって撃退される。トランクスと名乗るその青年は、自分は未来からやってきたブルマとベジータの息子であることを悟空にだけ明かすと同時に、悟空は心臓病によって命を落とすこと、3年後に現れる2体の人造人間が絶望の未来をもたらすことを告げる。その後本当に心臓病によって危篤状態になるも、トランクスから渡された特効薬によって生還、来るべき日に備えて3年間各々に修行してその日を迎える悟空たちであったが、事態はトランクスが知っている歴史とは大きく違うものとなり、彼さえ知らなかった人造人間たちまで現れ、さらに究極の人造人間・セルが未来から出現。悟空らの想定を遥かに超えた戦士が続々と現れた。
-        人造人間17号と人造人間18号を吸収することで完全体となったセルは地球の命運を賭けた武道会「セルゲーム」の開催を全世界に宣言する。悟空らは天界にある1日で1年の修行が行えるも過酷な環境に晒される「精神と時の部屋」で修行し、強さを増してセルゲームに臨むが、悟空はこのセルとの闘いの中で地球を守るために命を落とす。だが、その遺志を受け継いだ息子・悟飯が超サイヤ人2へと覚醒、父・悟空の幻影と共にかめはめ波を放ちセルを撃破。セルゲームを制した悟飯たちは、ドラゴンボールによりセルに殺された人々を蘇生させるが、悟空は自分が悪人を引き寄せているという理由で生き返りを拒否し、あの世に残ることを選ぶ。
-      </p>
-
-      <h1>魔人ブウ編</h1>
-
-      <p>
-        セルゲームより約7年後、高校生に成長した悟飯が天下一武道会に出場することを知った悟空は、自らも出場するために占いババの力によって1日だけこの世に戻る。天下一武道会の最中、悟空たちは界王よりもさらに高位の存在である界王神から、恐ろしい力を持つ魔人ブウの封印が解かれようとしていることを知らされる。復活した魔人ブウにより悟飯やベジータが倒され、悟空はあの世に帰ったため、地球の命運は悟空の次男・孫悟天と少年トランクスの幼い二人に託される。
-        一方、魔人ブウは様々な人間との出会いからより邪悪で強力な魔人へと変貌。悟天とトランクスが「フュージョン（融合）」して誕生した戦士・ゴテンクスや、潜在能力を解放し、パワーアップを遂げて帰ってきた悟飯らが応戦するが、戦士たちを次々と吸収し姿を変えていく魔人ブウに苦戦を強いられる。危機に陥った悟飯らを救うため現世に舞い戻った悟空とベジータは、界王神界で真の姿となった魔人ブウとの最終決戦に臨む。ドラゴンボールの協力もあり、地球・ナメック星・あの世の人々のエネルギーによって作り上げられた超特大の元気玉によって魔人ブウは完全に消滅する。
-        それから10年後、悟空は孫のパンと共に天下一武道会に久しぶりに出場し、魔人ブウの生まれ変わりである少年・ウーブと出会う。悟空はウーブと共に見果てぬ強さを追い求めて修行に旅立ち、物語は幕を閉じる。
-      </p>
     </>
   )
 }
 
 describe("<ScrollArea />", () => {
-  test("renders component correctly", async () => {
-    await a11y(
-      <ScrollArea h="xs">
-        <TestContent />
-      </ScrollArea>,
-    )
-  })
-
-  test("renders children content correctly", async () => {
-    await render(
-      <ScrollArea>
-        <p>Item 1</p>
-      </ScrollArea>,
-    )
-
-    await expect.element(page.getByText("Item 1")).toBeInTheDocument()
-  })
-
-  test("updates scroll position when controlled externally", async () => {
-    const { container } = await render(
-      <ScrollArea>
-        <TestContent />
-      </ScrollArea>,
-    )
-
-    setScrollTop(container, 0)
-    container.dispatchEvent(new Event("scroll", { bubbles: true }))
-    expect(container.scrollTop).toBe(0)
-
-    setScrollTop(container, 200)
-    container.dispatchEvent(new Event("scroll", { bubbles: true }))
-    expect(container.scrollTop).toBe(200)
-  })
-
-  test("calls onScrollPositionChange when scrolled", async () => {
-    const mockScrollPositionChange = vi.fn()
-
-    await render(
-      <ScrollArea
-        data-testid="scroll-area"
-        onScrollPositionChange={mockScrollPositionChange}
-      >
-        <TestContent />
-      </ScrollArea>,
-    )
-
-    const scrollArea = page
-      .getByTestId("scroll-area")
-      .element() as HTMLDivElement
-    setScrollTop(scrollArea, 100)
-    scrollArea.dispatchEvent(new Event("scroll", { bubbles: true }))
-
-    expect(mockScrollPositionChange).toHaveBeenCalledExactlyOnceWith({
-      x: 0,
-      y: 100,
-    })
-  })
-
   test("shows scroll indicators on hover and hides them on leave", async () => {
     const { user } = await render(
       <div style={{ alignItems: "flex-start", display: "flex", gap: "320px" }}>
@@ -166,7 +79,10 @@ describe("<ScrollArea />", () => {
     )
 
     const scrollArea = page.getByTestId("scroll-area")
-    const scrollAreaElement = scrollArea.element() as HTMLDivElement
+    const scrollAreaElement = scrollArea.element()
+    if (!(scrollAreaElement instanceof HTMLDivElement)) {
+      throw new Error("scroll-area is not an HTMLDivElement")
+    }
 
     await expect.element(scrollArea).toHaveAttribute("data-hidden")
     await expect.element(scrollArea).not.toHaveAttribute("data-scroll")
@@ -186,7 +102,6 @@ describe("<ScrollArea />", () => {
     const firstScrollCallCount = onScrollPositionChange.mock.calls.length
     await expect.element(scrollArea).not.toHaveAttribute("data-hidden")
 
-    // Scroll again to trigger clearTimeout (line 95) before the previous timeout fires
     scrollAreaElement.scrollTop =
       scrollAreaElement.scrollHeight - scrollAreaElement.clientHeight
     scrollAreaElement.dispatchEvent(new Event("scroll", { bubbles: true }))
@@ -202,166 +117,5 @@ describe("<ScrollArea />", () => {
     await expect.element(scrollArea).not.toHaveAttribute("data-hidden")
     await expect.element(scrollArea).not.toHaveAttribute("data-scroll")
     await expect.element(scrollArea).toHaveAttribute("data-hidden")
-  })
-
-  test("applies safari specific key format", async () => {
-    const originalPlatform = Object.getOwnPropertyDescriptor(
-      window.navigator,
-      "platform",
-    )
-    const originalVendor = Object.getOwnPropertyDescriptor(
-      window.navigator,
-      "vendor",
-    )
-    const originalUserAgentData = Object.getOwnPropertyDescriptor(
-      window.navigator,
-      "userAgentData",
-    )
-
-    // Mock Safari environment
-    Object.defineProperty(window.navigator, "platform", {
-      configurable: true,
-      value: "MacIntel",
-    })
-    Object.defineProperty(window.navigator, "vendor", {
-      configurable: true,
-      value: "Apple Computer, Inc.",
-    })
-    Object.defineProperty(window.navigator, "userAgentData", {
-      configurable: true,
-      value: { platform: "macOS" },
-    })
-
-    try {
-      await render(
-        <ScrollArea type="never" data-testid="scroll-area">
-          <TestContent />
-        </ScrollArea>,
-      )
-
-      const scrollArea = page.getByTestId("scroll-area").element()
-
-      expect(scrollArea).toHaveAttribute("data-key")
-
-      // The key should end with the expected pattern
-      expect(scrollArea.getAttribute("data-key")).toMatch(/-false-false$/)
-    } finally {
-      if (originalPlatform) {
-        Object.defineProperty(window.navigator, "platform", originalPlatform)
-      } else {
-        Reflect.deleteProperty(window.navigator, "platform")
-      }
-
-      if (originalVendor) {
-        Object.defineProperty(window.navigator, "vendor", originalVendor)
-      } else {
-        Reflect.deleteProperty(window.navigator, "vendor")
-      }
-
-      if (originalUserAgentData) {
-        Object.defineProperty(
-          window.navigator,
-          "userAgentData",
-          originalUserAgentData,
-        )
-      } else {
-        Reflect.deleteProperty(window.navigator, "userAgentData")
-      }
-    }
-  })
-
-  test("merges consumer root props and preserves event handler order", async () => {
-    const order: string[] = []
-    const restOnMouseEnter = vi.fn(() => order.push("rest-enter"))
-    const restOnMouseLeave = vi.fn(() => order.push("rest-leave"))
-    const restOnScroll = vi.fn(() => order.push("rest-scroll"))
-    const propsOnMouseEnter = vi.fn(() => order.push("props-enter"))
-    const propsOnMouseLeave = vi.fn(() => order.push("props-leave"))
-    const propsOnScroll = vi.fn(() => order.push("props-scroll"))
-    const onScrollPositionChange = vi.fn(() => order.push("internal-scroll"))
-
-    const { result } = await renderHook(() =>
-      useScrollArea({
-        className: "rest-class",
-        style: { color: "tomato" },
-        onMouseEnter: restOnMouseEnter,
-        onMouseLeave: restOnMouseLeave,
-        onScroll: restOnScroll,
-        onScrollPositionChange,
-      }),
-    )
-
-    const rootProps = result.current.getRootProps({
-      className: "props-class",
-      style: { background: "black" },
-      onMouseEnter: propsOnMouseEnter,
-      onMouseLeave: propsOnMouseLeave,
-      onScroll: propsOnScroll,
-    })
-
-    expect(rootProps.className).toContain("rest-class")
-    expect(rootProps.className).toContain("props-class")
-    expect(rootProps.style).toMatchObject({
-      background: "black",
-      color: "tomato",
-      overflow: "auto",
-    })
-
-    rootProps.onMouseEnter?.({} as any)
-    rootProps.onMouseLeave?.({} as any)
-    rootProps.onScroll?.({
-      target: { scrollLeft: 16, scrollTop: 24 },
-    } as any)
-
-    expect(order).toStrictEqual([
-      "rest-enter",
-      "props-enter",
-      "rest-leave",
-      "props-leave",
-      "rest-scroll",
-      "props-scroll",
-      "internal-scroll",
-    ])
-  })
-
-  test("composes refs once in safari root props", async () => {
-    Object.defineProperty(window.navigator, "platform", {
-      value: "MacOS",
-      writable: true,
-    })
-    Object.defineProperty(window.navigator, "vendor", {
-      value: "Apple Computer, Inc.",
-      writable: true,
-    })
-
-    const externalRef = vi.fn()
-    const consumerRef = vi.fn()
-
-    const Component = () => {
-      const { getRootProps } = useScrollArea({
-        ref: externalRef,
-      })
-
-      return (
-        <div
-          data-testid="scroll-area-ref"
-          {...getRootProps({ ref: consumerRef })}
-        />
-      )
-    }
-
-    await render(<Component />)
-
-    const scrollArea = page.getByTestId("scroll-area-ref").element()
-
-    expect(scrollArea).toHaveAttribute("data-key")
-    expect(externalRef).toHaveBeenCalledWith(scrollArea)
-    expect(consumerRef).toHaveBeenCalledWith(scrollArea)
-    expect(
-      externalRef.mock.calls.filter(([value]) => value === scrollArea),
-    ).toHaveLength(1)
-    expect(
-      consumerRef.mock.calls.filter(([value]) => value === scrollArea),
-    ).toHaveLength(1)
   })
 })
